requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.5.0
jinja2>=3.1.2
pyyaml>=6.0.1
python-dotenv>=1.0.0

# PDF generation
reportlab==4.0.0
PyPDF2==3.0.0
pdfplumber==0.9.0
qrcode[pil]==7.4.2

# Web automation
playwright==1.40.0
beautifulsoup4==4.12.0
requests==2.31.0
selenium==4.15.0

# Data processing
pandas==2.0.0
openpyxl==3.1.0
gspread==5.12.0
google-auth==2.25.0

# Video processing
moviepy==1.0.3
pillow==10.0.0

# API integrations
gumroad-api==0.1.0
stripe==7.0.0
mailchimp3==3.0.0

# Task scheduling
schedule==1.2.0
celery==5.3.0
redis==5.0.0

# FastAPI and web dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.25.2

# Utilities
click==8.1.0
rich==13.7.0
tqdm==4.66.0
